+++ generators/app/templates/02-microservice/src/app.service.ts
@@ -1,0 +1,23 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import * as packageInfo from '../package.json';
+
+@Injectable()
+export class AppService {
+  constructor(protected readonly config: ConfigService) {}
+
+  getHealthCheck(): string {
+    return 'OK!';
+  }
+
+  apiInfo() {
+    return {
+      name: '<%= appName %>',
+      version: packageInfo.version,
+      location: {
+        docs: '/asyncapi',
+        healthCheck: '/health',
+      },
+    };
+  }
+}