+++ generators/app/templates/02-microservice/src/common/constants.ts
@@ -1,0 +1,9 @@
+<% if (usePubSub) { %>
+export const PUBSUB_SERVICE = 'PUBSUB_SERVICE';
+<% } %>
+<% if (useKafka) { %>
+export const KAFKA_SERVICE = 'KAFKA_SERVICE';
+<% } %>
+<% if (useRabbitmq) { %>
+export const RABBITMQ_SERVICE = 'RABBITMQ_SERVICE';
+<% } %>