+++ generators/module/templates/01-basic/schemas/__moduleName__.schema.ts
@@ -1,3 +1,4 @@
+import { ApiProperty } from '@nestjs/swagger';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Document } from 'mongoose';
 
@@ -6,6 +7,7 @@ export type <%= kebabToPascal(config.name) %>Document = <%= kebabToPascal(config
 @Schema()
 export class <%= kebabToPascal(config.name) %> {
   @Prop()
+  @ApiProperty({ description: 'The name of the <%= kebabToPascal(config.name) %>' })
   name: string;
 }