+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Change Log
+## 1.1.0
+- Add project type `microservice`
+- Update use openapi/asyncapi for API Document (solution for both of standalone app & micoservice)
+- Bugs fixing
 ## 1.0.1
 - Fix issue initial project with GCP Pubsub option